move System.Drawing to .Net Framework DNNrocketAPI

+++ API/Componants/Login.cs
@@ -92,10 +92,13 @@ namespace DNNrocketAPI.Componants
 
         public static string LoginForm(SimplisityInfo systemInfo, SimplisityInfo sInfo, string interfacekey,int userid)
         {
-            // clear cookie for cmd.  This could cause a fail login loop.
-            // A module MUST always have a tabid and a valid users.  Invalid cookies without tabid could cause a loop.
-            DNNrocketUtils.DeleteCookieValue("s-cmd-menu-" + systemInfo.GUIDKey);
-            DNNrocketUtils.DeleteCookieValue("s-fields-menu-" + systemInfo.GUIDKey);
+            if (systemInfo != null)
+            {
+                // clear cookie for cmd.  This could cause a fail login loop.
+                // A module MUST always have a tabid and a valid users.  Invalid cookies without tabid could cause a loop.
+                DNNrocketUtils.DeleteCookieValue("s-cmd-menu-" + systemInfo.GUIDKey);
+                DNNrocketUtils.DeleteCookieValue("s-fields-menu-" + systemInfo.GUIDKey);
+            }
 
             if (userid > 0)
             {
